Change $my_sql* for {{ my_sql* }} in order to comply with ansible 1.2+
Otherwise, the following message appears (in case of ansible 1.4):

 [DEPRECATION WARNING]: Legacy variable subsitution, such as using ${foo} or
$foo instead of {{ foo }} is currently valid but will be phased out and has
been out of favor since version 1.2. This is the last of legacy features on our
deprecation list. You may continue to use this if you have specific needs for
now. This feature will be removed in version 1.6. Deprecation warnings can be
disabled by setting deprecation_warnings=False in ansible.cfg.

Signed-off-by: Rogerio Prado de Jesus <rogeriopradoj@gmail.com>